rename and simplify command
