+++ migration-engine/core/src/api.rs
@@ -43,7 +43,7 @@ where
 // liking them in the exported class.
 #[async_trait::async_trait]
 pub trait GenericApi: Send + Sync + 'static {
-    async fn version(&self, input: &VersionInput) -> CoreResult<String>;
+    async fn version(&self, input: &serde_json::Value) -> CoreResult<String>;
     async fn apply_migration(&self, input: &ApplyMigrationInput) -> CoreResult<MigrationStepsResultOutput>;
     async fn apply_migrations(&self, input: &ApplyMigrationsInput) -> CoreResult<ApplyMigrationsOutput>;
     async fn calculate_database_steps(
@@ -83,8 +83,8 @@ where
     C: MigrationConnector<DatabaseMigration = D>,
     D: DatabaseMigrationMarker + Send + Sync + 'static,
 {
-    async fn version(&self, input: &VersionInput) -> CoreResult<String> {
-        self.handle_command::<VersionCommand<'_>>(input)
+    async fn version(&self, input: &serde_json::Value) -> CoreResult<String> {
+        self.handle_command::<VersionCommand>(input)
             .instrument(tracing::info_span!("Version"))
             .await
     }
